All tests pass
git-svn-id: https://svn.smr.arpa.emr.it/arkimet/arkimet/trunk@2227 9b5092fc-ae44-42f6-b796-0d08dad2e17b

+++ arki/emitter/json.h
@@ -60,7 +60,7 @@ public:
 
     virtual void add_null();
     virtual void add_bool(bool val);
-    virtual void add_int(int val);
+    virtual void add_int(long long int val);
     virtual void add_double(double val);
     virtual void add_string(const std::string& val);
